Refactor common folder
