+++ client/go.mod
@@ -6,7 +6,7 @@ require (
 	github.com/briandowns/spinner v1.19.0
 	github.com/creack/pty v1.1.18
 	github.com/google/uuid v1.3.0
-	github.com/runopsio/hoop/proto v0.0.0-00010101000000-000000000000
+	github.com/runopsio/hoop/common v0.0.0-00010101000000-000000000000
 	github.com/spf13/cobra v1.5.0
 	golang.org/x/term v0.0.0-20220919170432-7a66f970e087
 	google.golang.org/grpc v1.49.0
@@ -26,4 +26,4 @@ require (
 	google.golang.org/protobuf v1.28.1 // indirect
 )
 
-replace github.com/runopsio/hoop/proto => ../proto
+replace github.com/runopsio/hoop/common => ../common
