+++ gateway/agent/api.go
@@ -1,10 +1,11 @@
 package agent
 
 import (
+	"net/http"
+
 	"github.com/gin-gonic/gin"
 	"github.com/google/uuid"
 	"github.com/runopsio/hoop/gateway/user"
-	"net/http"
 )
 
 type (
