+++ gateway/api/middleware.go
@@ -1,11 +1,14 @@
 package api
 
 import (
-	"github.com/gin-gonic/gin"
-	"github.com/golang-jwt/jwt/v4"
-	pb "github.com/runopsio/hoop/proto"
+	"fmt"
+	"log"
 	"net/http"
 	"strings"
+
+	"github.com/gin-gonic/gin"
+	"github.com/golang-jwt/jwt/v4"
+	pb "github.com/runopsio/hoop/common/proto"
 )
 
 func CORSMiddleware() gin.HandlerFunc {
@@ -25,46 +28,47 @@ func CORSMiddleware() gin.HandlerFunc {
 }
 
 func (api *Api) Authenticate(c *gin.Context) {
-	email := parseEmail(c)
+	email, err := parseEmail(c)
+	if err != nil {
+		log.Printf("failed authenticating, err=%v", err)
+		c.AbortWithStatus(http.StatusUnauthorized)
+		return
+	}
 
 	ctx, err := api.UserHandler.Service.ContextByEmail(email)
 	if err != nil {
 		c.AbortWithStatus(http.StatusUnauthorized)
+		return
 	}
 
 	if ctx == nil {
 		c.AbortWithStatus(http.StatusUnauthorized)
+		return
 	}
 
 	c.Set("context", ctx)
 	c.Next()
 }
 
-func parseEmail(c *gin.Context) string {
+func parseEmail(c *gin.Context) (string, error) {
 	if PROFILE == pb.DevProfile {
-		return "tester@hoop.dev"
+		return "tester@hoop.dev", nil
 	}
 
 	tokenHeader := c.GetHeader("authorization")
-
 	tokenParts := strings.Split(tokenHeader, " ")
 	if len(tokenParts) != 2 || tokenParts[0] != "Bearer" || tokenParts[1] == "" {
-		c.AbortWithStatus(http.StatusUnauthorized)
+		return "", fmt.Errorf("failed parsing token from header")
 	}
-
 	tokenValue := tokenParts[1]
 
 	token, err := jwt.Parse(tokenValue, jwks.Keyfunc)
 	if err != nil {
-		c.AbortWithStatus(http.StatusUnauthorized)
+		return "", err
 	}
 
-	var email string
 	if claims, ok := token.Claims.(jwt.MapClaims); ok && token.Valid {
-		email = claims["https://runops.io/email"].(string)
-	} else {
-		c.AbortWithStatus(http.StatusUnauthorized)
+		return claims["https://runops.io/email"].(string), nil
 	}
-
-	return email
+	return "", fmt.Errorf("invalid token")
 }
